+++ docs/source/conf.py
@@ -18,11 +18,11 @@
 # -- Project information -----------------------------------------------------
 
 project = 'nestli'
-copyright = '2022, Aaron Bojarski, Fazel Khayatian, Hanmin Cai'
+copyright = '2023, Aaron Bojarski, Fazel Khayatian, Hanmin Cai'
 author = 'Aaron Bojarski, Fazel Khayatian, Hanmin Cai'
 
 # The full version, including alpha/beta/rc tags
-release = '1.0.0'
+release = '1.1.0'
 
 
 # -- General configuration ---------------------------------------------------