+++ src/nestli/manager/__init__.py
@@ -6,6 +6,7 @@ MOSAIK_CONFIG = {
     "CONSTANT_VALUE": {"python": "nestli.simulators.ConstantValue:ConstantValue"},
     "DATABASE_DATA": {"python": "nestli.simulators.DataBaseData:DataBaseData"},
     "PYTHON_FUNCTION": {"python": "nestli.simulators.PythonFunction:PythonFunction"},
+    "DATABASE_DATA_FORECAST": {"python": "nestli.simulators.DataBaseDataForecast:DataBaseDataForecast"},
 }
 
 DATA_DOWNLOAD = "https://polybox.ethz.ch/index.php/s/QgaDRyA2oXoFs1C/download"