+++ docs/source/releasenotes.rst
@@ -3,6 +3,20 @@ Release Notes
 ==============
 
 
+1.1.0
+-------
+- update mosaik to resolve bug with simulator simulation order
+- move other source files to zip
+- minor changes in template
+
+
+1.0.0
+-------
+- first official release of nestli
+- added documentation
+- template has more inputs
+
+
 0.2.0
 -------
 - added full python controller functionality and template